% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubular.R
\name{PlotTubecolours}
\alias{PlotTubecolours}
\title{PlotTubecolours}
\usage{
PlotTubecolours(extended = FALSE)
}
\arguments{
\item{extended}{(FALSE) additionally include non-tube lines on the underground map}
}
\value{
A ggplot of the colours of TubeColours()
}
\description{
PlotTubecolours
}
\examples{
PlotTubecolours(extended=TRUE)
}
