% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubular.R
\name{TubeColours}
\alias{TubeColours}
\title{TubeColours}
\usage{
TubeColours(extended = FALSE)
}
\arguments{
\item{extended}{(FALSE) additionally return non-tube lines on the underground map}
}
\value{
A dataframe of colour values and line names
}
\description{
TubeColours
}
\examples{
TubeColours()
}
